#20: Fixing Bug in ISO TP on MCU Side + UDS Comm Spec

+++ MCU_Aurix/bootloader/inc/isotp.h
@@ -42,7 +42,7 @@ typedef struct isoTp_RX
     uint8_t* data;
     uint8_t* write_ptr;
 
-    uint16_t data_in_len;
+    uint32_t data_in_len;
     uint8_t ready_to_read; // bool that will be set to =! 0 if message can be read.
     uint8_t last_consecutive_ctr;
